feat: добавить регистрацию и авторизацию

+++ controllers/cards.js
@@ -1,39 +1,44 @@
 const Card = require('../models/card');
-const {
-  sendBadRequest,
-  sendCardNotFound,
-  sendServerError,
-  sendBadId,
-} = require('../utils/errors');
+const { notFoundError, forbiddenError } = require('../utils/errors');
 
-module.exports.getAllCards = (req, res) => {
+module.exports.getAllCards = (req, res, next) => {
   Card.find({})
-    .then((cards) => res.send(cards))
-    .catch(() => sendServerError(res));
+    .then((cards) => (cards.length > 0 ? res.send(cards) : Promise.reject(notFoundError('Список карточек пуст'))))
+    .catch(next);
 };
 
-module.exports.addCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const owner = req.user._id;
   const { name, link } = req.body;
   Card.create({ name, link, owner })
     .then((card) => res.send(card))
-    .catch((err) => (err.name === 'ValidationError' ? sendBadRequest(res) : sendServerError(res)));
+    .catch(next);
 };
 
-module.exports.deleteCardById = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .then((card) => (card ? res.send(card) : sendCardNotFound(res)))
-    .catch((err) => (err.name === 'CastError' ? sendBadId(res) : sendServerError(res)));
+module.exports.deleteCardById = (req, res, next) => {
+  Card.findById(req.params.cardId)
+    .then((card) => {
+      if (!card) {
+        throw notFoundError('Карточка не найдена');
+      }
+      if (card.owner.toString() !== req.user._id) {
+        throw forbiddenError('Ошибка доступа');
+      }
+
+      return Card.findByIdAndRemove(req.params.cardId)
+        .then((validCard) => (validCard && res.send(validCard)));
+    })
+    .catch(next);
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $addToSet: { likes: req.user._id } }, { new: true })
-    .then((card) => (card ? res.send(card) : sendCardNotFound(res)))
-    .catch((err) => (err.name === 'CastError' ? sendBadRequest(res) : sendServerError(res)));
+    .then((card) => (card ? res.send(card) : Promise.reject(notFoundError('Карточка не найдена'))))
+    .catch(next);
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   Card.findByIdAndUpdate(req.params.cardId, { $pull: { likes: req.user._id } }, { new: true })
-    .then((card) => (card ? res.send(card) : sendCardNotFound(res)))
-    .catch((err) => (err.name === 'CastError' ? sendBadRequest(res) : sendServerError(res)));
+    .then((card) => (card ? res.send(card) : Promise.reject(notFoundError('Карточка не найдена'))))
+    .catch(next);
 };
